Adds animal and food prefabs
Adds prefabs for various animal types and food.
Also renames `PlayerController` to `DestroyOutOfBounds`
and modifies its behavior to destroy objects when they
move out of bounds and creates a new script `MoveForward`
for moving spawned objects forward.
The `PlayerController` script is modified to handle player
movement and spawning of food objects.
A new scene "Unit 2, Version 1.2" with the created
assets and updated scripts is created.